ipo-scheduler/requirements.txt
fastapi 
uvicorn 
beanie 
motor
pytest
pyjwt
pymongo
python-dotenv
concurrent-log-handler
jinja2
python-jose
passlib
pytest-asyncio
korean-lunar-calendar
apscheduler
requests
beautifulsoup4
pydantic[email]
pandas
scikit-learn
sympy
chromedriver-autoinstaller
selenium
